refactor(lnd): rename LND components to Lightning
