+++ src/components/designer/lightning/InfoTab.tsx
