+++ src/components/designer/lightning/actions/RemoveNode.spec.tsx
