+++ src/components/designer/lightning/actions/RemoveNode.tsx
