+++ src/components/designer/lightning/connect/EncodedStrings.tsx
